+++ src/wreqr.commands.js
@@ -7,11 +7,37 @@ Wreqr.Commands = (function(Wreqr){
   "use strict";
 
   return Wreqr.Handlers.extend({
-    execute: function(){
-      var name = arguments[0];
-      var args = Array.prototype.slice.call(arguments, 1);
+    constructor: function(){
+      this.storage = new Wreqr.CommandStorage();
+      this.on("handler:add", this.executeCommands, this);
 
-      this.getHandler(name).apply(this, args);
+      var args = Array.prototype.slice.call(arguments);
+      Wreqr.Handlers.prototype.constructor.apply(this, args);
+    },
+
+    // Execute a named command with the supplied args
+    execute: function(name, args){
+      name = arguments[0];
+      args = Array.prototype.slice.call(arguments, 1);
+
+      if (this.hasHandler(name)){
+        this.getHandler(name).apply(this, args);
+      } else {
+        this.storage.addCommand(name, args);
+      }
+
+    },
+
+    // Internal method to handle bulk execution of stored commands
+    executeCommands: function(name, handler, context){
+      var command = this.storage.getCommands(name);
+
+      // loop through and execute all the stored command instances
+      _.each(command.instances, function(args){
+        handler.apply(context, args);
+      });
+
+      this.storage.clearCommands(name);
     }
   });