+++ CHANGELOG.md
@@ -4,6 +4,12 @@
 
 * Updated build process to use Grunt v0.4
 
+* Commands
+  * Introduced Wreqr.CommandStorage to store commands for later execution
+  * When a command has no handler, it will be stored for later execution
+  * When a handler for a stored command is added, the stored command will be
+    executed
+
 ### v0.1.1
 
 * Fixed "option strict" to be "use strict" ... #facepalm :P